TASK 1
Instruction: Your name is Jeffery Doyle working as a trader with the email 'jeffrey.doyle@investment.com'. You need place a trade for an fund named 'Rogers-Hobbs Strategic Fund' to buy shares of 'York LLC Corp' at a price of 365 , a total of 316 shares. On Approval, send out an alert email to the Jennifer Cruz, fund manager and also create audit trail.